% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_did_causal_forests.R
\name{estimate_DiDCF}
\alias{estimate_DiDCF}
\title{Estimate Difference-in-Difference Causal Forests for Long Format Data}
\usage{
estimate_DiDCF(Y, W, X, b, time, id, ...)
}
\arguments{
\item{Y}{A numeric vector of the outcome variable.}

\item{W}{A vector of treatment indicators (one per unit).}

\item{X}{A numeric matrix or data frame of features that vary only in the cross-sectional dimension (one row per unit).}

\item{b}{The value in the \code{time} vector corresponding to the base period.}

\item{time}{A vector indicating the time period for each observation in Y.}

\item{id}{A vector indicating the unit identifier for each observation in Y.}

\item{...}{Additional arguments to be forwarded to \code{grf::causal_forest}.}
}
\value{
A list of causal forest objects (one for each time period other than the base period).
}
\description{
This function computes the difference outcome
\deqn{Z_{it} = Y_{it} - Y_{ib}}
for each time period \( t \) (using base period \( b \)) and estimates a causal forest
using the grf package on \((Z, W, X)\). Y is assumed to be in long format. Therefore, you
must provide the corresponding \code{time} and \code{id} vectors. It is required that every unit
(as identified by \code{id}) has an observation in the base period \( b \).
}
\details{
Additional arguments (...) are passed directly to the call to \code{grf::causal_forest}.
}
\examples{
estimate_DiDCF(example_data$Y,
               example_data$t_indicator[example_data$period==1],
               as.data.frame(model.matrix(~.,data=example_data[example_data$period==1,c("x_1","x_2")])),
               1,
               data$period,
               data$unit_id)
}
