% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Example Data}
\format{
A data frame with 12800 rows and 11 variables:
\describe{
\item{unit_id}{A numeric vector representing the ID of each worker.}
\item{firm_id}{A numeric vector representing the ID of each firm.}
\item{period}{A numerric vector indicating time period 1 or 2}
\item{x_1}{A factor vector indicating 1 or 0 on a given characteristic x_1}
\item{x_2}{A factor vector indicating 1 or 0 on a given characteristic x_2}
\item{firm_variance}{firm-specific variance}
\item{treated}{a vector indicating treatment in the second period, time and cross-sectional variation}
\item{treat_prob}{a vector indicating probability of treatment}
\item{t_indicator}{a vector indicating treated units, cross-sectional variation}
\item{Y}{outcome variable}
}
}
\usage{
example_data
}
\description{
A dataset containing example data from the Simulation section of
Gavrilova, Langoergen and Zoutman (2025)
Difference-in-Difference Causal Forests with an Application to Payroll Tax Incidence in Norway
}
\examples{
example_data

}
\keyword{datasets}
