% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vip.R
\name{vip}
\alias{vip}
\title{Title}
\usage{
vip(forest)
}
\arguments{
\item{forest}{GRF forest object of type causal_forest, a causal forest estimate}
}
\value{
A table with two columns, variable names and values of variables importance
}
\description{
Title
}
\examples{
my_table<-vip(my_forest)

}
